Initial version of proposal for the new riscv transport and adapter
Separate makefile for transport, no autoconf changes with riscv flag

Initial autoconf changes

socket transport appears in the transport list command

Renamed socket to be more relatable to riscv architecture, for DMI, which is riscv specific thing

Improved socket, added host and port parsing and selection of transport for the adapter

Remove unnecessary include when it is done in header

Got rid of the hacky way of making the compiler link the new socket transport lib, doing it in adapter init instead, for every transport layer it supports.

Getting around JTAG (taps, chain-positions and cfg arguments meant for it)

Prevent a driver to get initialized twice

Bypassing the dtmcontrol_scan for jtag with socket

Localhost did not work, but the ip address for HOME does, socket can try to send something to server now

Wrong bits extracted, 1st byte was a read/write command, there was an offset here

Adding so far's python server simulator to make socket testing easier for socket transport layer

Made some fixes of some ptr bugs and how we advance in recv_all, but the issue still remains: data_buffer is all zeroes, examine that

Some other bugs fixed and out of bounds and fixed the shift

Improved server a bit by removing some unnecessary parts that sent the response twice and adding more logging/prints in the code

Adding no timeout and wait indefinitely (for send and recv) and improve the code a bit. Still getting all zeroes, though

Fix the server for the next command and adding more jtag guarded checks so that it could work via the socket.

Further adding the checks around JTAB and reset buffer in server code

Going around the JTAG buffered dmi_read for the socket

Now the server and riscv-openocd communicate constantly, although server is sending the same data to riscv-openocd

Returning the right value according to openocd's needs

A bit more improved server to align expectations what openocd expects

Bad indentation: infinite loop

Fixing the method to send a request to the target via socket and not to use the JTAG construct

Improved logging, dmi_write to have the socket version, too and some changed on the server to be more responsive

Listening on the port achieved, but with an ugly hack of temporary server reply

Adding an alternative of partially received buffer on the server side, adjust it to reach the same point as main1

Making sure now the server clears up the old data and with each received command starts anew

More precise assigning of data according to the riscv spec on https://riscv.org/wp-content/uploads/2024/12/riscv-debug-release.pdf pagde 21

Passing the stage of examining target, up to the stage where hart reset request is set

Commenting out troublesomecode to continue initialization

Some attempt of changes on OpenOCD's side and trying to remove some (maybe) not needed parts, also improving a server (python) for abstract commands

Some more crash fixes and restoring some commented out code that was actually needed for some future computations, otherwise there would have been a crash

Adding the read part in riscv batch dm read to be able to keep the counter to non-zero so that the future assertion passes

Adjusting the python server for reg number 0x300 needed from openocd's side

Potential memory corruption with this log line here

Putting back the log.

The error was not in log but rather in how log interprets the data and how prntf does it, that is more resilient.

The issue was in allocated and uninitialized, that caused some issues later on. This is why I used calloc for now.

With this approach it seems to work and to give positive results.

abstract_data_get_from_batch was needed to set the value of dcsr not to be uninitialized

Adding a counter so that the next time handle_dmi_read is called it returns another value, just like the spike simulator

Changing the server so that on the 2nd and future tries returns the same value.

Handling DM_STATUS_ALLHALTED later in the code execution to pass the halting stage on the server side

Adding risv guard around one more place that has it, that was skipped for socket, later in the execution logic

Uncommenting commented out parts for which I thought they cause issues

Server update with commented out support for gdb rsp packages and removed code that is not used plus some logging added

Unnecessary check when it was done in dm_read for platform

Changing the approach and moving the riscv_socket_dmi files to jtag, close to the core.h and core.c files so that a cleaner wrapper could be made

Wrapper for socket and jtag and other future transport layers added, but not yet used yet in the rest of the code. Replace jtag calls with wrapper calls.

Macro and not a static function expansion of reusable check due to the short circuiting and evaluation of parameters when passed to a function as opposed to the macro, where tap is a problematic access

Using wrapper for jtag in batch.c and fixing the issue with premature endinf og ifdef for file defined in riscv_socket_dmi.h

Changing macros in final places of riscv-013.c

Reset bit coming as 0, not as 1 anymore

Some pieces combined to be deeper decided if it is JTAG or socket (or other transport)

Revert "Some pieces combined to be deeper decided if it is JTAG or socket (or other transport)"

This reverts commit 46ddbebf59f9842b9d7d714758a96a5088670cd0.

Some pieces combined to be deeper decided if it is JTAG or socket (or other transport)

Dtm control scan reused and moved check deeper inside the function stack

dtmcontrol_scan also unified to do a check deeper in the functions it is calling

Final touches of moving the JTAG check deeper as in dmi_write as in dmi_read

Server should handle the case for replying bask to DM_ABSTRACTCS  with the value as Spike simulator does

Bringing back malloc where it was

MISA gets the value different than zero now for HW capabilities

Update misa access via socket by completely avoiding JTAG and buffers

Trying to use more riscv compliant server simulators to test socket dmi openocd connection to this target

more modern version of python server and openocd to accommodate communication with it