Putting back the log.
The error was not in log but rather in how log interprets the data and how prntf does it, that is more resilient.

The issue was in allocated and uninitialized, that caused some issues later on. This is why I used calloc for now.

With this approach it seems to work and to give positive results.